issue-723 fix sql scripts: use integer column type.

+++ src/main/resources/db/migration/oracle/V2.7.7.20200430193000__add_amount_files_in_analysis_result.sql
@@ -1,1 +1,1 @@
-ALTER TABLE ${ohdsiSchema}.ee_analysis_status ADD amount_files_in_analysis NUMBER(19);
+ALTER TABLE ${ohdsiSchema}.ee_analysis_status ADD amount_files_in_analysis INTEGER;
